Create MessageManager class and moved related function from BambooAI class to MessageManager class
